Rename util.py -> arg_check.py
more descriptive name.

Also made a deploy path optional for the builder, since for KBase-style
configuration, we expect and env var.
