+++ src/jgikbase/idmapping/userhandlers/kbase_user_handler.py
@@ -2,7 +2,7 @@
 A ID mapper service user handler for KBase (https://kbase.us) user accounts.
 """
 from jgikbase.idmapping.core.user_handler import UserHandler
-from jgikbase.idmapping.core.util import not_none
+from jgikbase.idmapping.core.arg_check import not_none
 from jgikbase.idmapping.core.user import AuthsourceID, User, Username
 from jgikbase.idmapping.core.tokens import Token
 import requests
